pyproject.toml
[tool.poetry]
name = "fil2md"
version = "0.1.0"
description = "A web-based file converter that transforms various file formats into Markdown"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = ">=3.11,<3.14"
flask = "^3.1.0"
markitdown = {version = "^0.1.4", extras = ["all"]}
werkzeug = "^3.1.3"
openpyxl = "^3.1.5"
python-pptx = "^1.0.2"
pillow = "^11.0.0"
beautifulsoup4 = "^4.12.3"
pypdf2 = "^3.0.1"
lxml = "^5.3.0"
markdown = "^3.7"
python-docx = "^1.1.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
